pull

+++ src/components/utils/BtnAddFav.jsx
@@ -1,9 +1,14 @@
-import React, {useEffect, useState} from 'react'
-import {HiBookmark} from 'react-icons/hi'
-import {useAddNewFavoriteMutation, useDeleteFavoriteMutation} from '../../services/RTK/favoritesApi'
-import {dispatchAlert} from '../../helpers/alert'
+import React, { useEffect, useState } from 'react'
+import { HiBookmark } from 'react-icons/hi'
+import { useAddNewFavoriteMutation, useDeleteFavoriteMutation } from '../../services/RTK/favoritesApi'
+import { dispatchAlert } from '../../helpers/alert'
+import { useSelector } from 'react-redux'
+import swal from 'sweetalert'
+import { useNavigate } from 'react-router-dom'
 
-const BtnAddFav = ({favoriteStatus, gameId, userId}) => {
+const BtnAddFav = ({ favoriteStatus, gameId, userId }) => {
+    const isAuth = useSelector(state => state.auth.isAuth)
+    const nav = useNavigate()
     const [fav, setFav] = useState(null)
     const [addFavorite, state] = useAddNewFavoriteMutation()
     const [deleteFavorite, state2] = useDeleteFavoriteMutation()
@@ -14,7 +19,7 @@ const BtnAddFav = ({favoriteStatus, gameId, userId}) => {
 
     const createNewFavorite = () => {
         if (gameId && userId) {
-            addFavorite({gameId, userId})
+            addFavorite({ gameId, userId })
         }
         if (state.isError) {
             dispatchAlert('danger', 'Произошла ошибка')
@@ -23,7 +28,7 @@ const BtnAddFav = ({favoriteStatus, gameId, userId}) => {
 
     const deleteFav = () => {
         if (gameId && userId) {
-            deleteFavorite({gameId, userId})
+            deleteFavorite({ gameId, userId })
         }
 
         if (state2.isError) {
@@ -35,8 +40,13 @@ const BtnAddFav = ({favoriteStatus, gameId, userId}) => {
         <button
             type="button"
             onClick={() => {
-                setFav(!fav)
-                fav ? deleteFav() : createNewFavorite()
+                if (isAuth) {
+                    setFav(!fav)
+                    fav ? deleteFav() : createNewFavorite()
+                } else {
+                    swal('Пожалуйста, войдите или зарегистрируйтесь', { buttons: ['Отмена', 'Хорошо'] })
+                        .then((o) => o && nav('/login'))
+                }
             }}
             className={fav ? 'add-fav active' : 'add-fav'}
         >
